+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.0.0
 * Move default configuration path to /etc/linkctl
 * Support the Description attribute for netdevs
-* Add -v option to print version info
+* Add -version option to print version info
 * Improve UX of systemd-networkd restarts
 * Print an error if an unknown command is provided
-* Allow -a and -q options to default list command
+* Allow -a and -t options to default list command